requirements.txt
discord.py>=2.3.2
Pillow>=10.0.0
motor>=3.3.1
aiohttp>=3.9.1
python-dotenv>=1.0.0
dnspython>=2.4.2